style(msvc): fix C4244/C4267 in OptionRow/JsonUtil/GraphDisplay/AdjustSync/ActorScroller/DynamicActorScroller/Actor (item 2)
All static_cast, documenting existing intentional narrowing -- no
logic changes. sm_tests unchanged at 5966/226.

- OptionRow.cpp (4): a size_t course-entry count into int; a
  std::min<unsigned int> call's second arg; a lua_pushnumber(size_t)
  accessor.
- JsonUtil.h (4): four near-identical root.resize(v.size()) sites
  across SerializeVectorObjects/SerializeVectorPointers (x2)/
  SerializeArrayValues template helpers, each needing
  Json::Value::ArrayIndex instead of size_t.
- GraphDisplay.cpp (4): a size_t vertex count into DrawQuads's int; a
  size_t/int division into a fan count; two theme-metric ints into
  float size members.
- AdjustSync.cpp (4): a size_t size snapshotted before/after a filter
  call into an int counter (both sides of the diff); a size_t into
  unsigned int; a FormatNumberAndSuffix(int) call fed a size_t loop
  index.
- ActorScroller.cpp (4) + DynamicActorScroller.cpp (2, bonus find
  while in the file): a size_t sub-actor count into int; two
  std::ceil(float) results into int; three wrap(int&, size_t) sites
  across both scroller files.
- Actor.cpp (4): two RString::Left calls fed size_t-minus arithmetic
  parsing command/message names; a uint64_t timer value into a
  generic_global_timer_update(float, ...) call; a
  lua_pushnumber(size_t) accessor.

+++ src/ActorScroller.cpp
@@ -46,7 +46,7 @@ ActorScroller::ActorScroller()
 
 void ActorScroller::Load2()
 {
-	m_iNumItems = m_SubActors.size();
+	m_iNumItems = static_cast<int>(m_SubActors.size());
 
 	Lua *L = LUA->Get();
 	for( unsigned i = 0; i < m_SubActors.size(); ++i )
@@ -264,8 +264,8 @@ void ActorScroller::PositionItemsAndDrawPrimitives( bool bDrawPrimitives )
 
 	float fFirstItemToDraw = m_fCurrentItem - fNumItemsToDraw/2.f;
 	float fLastItemToDraw = m_fCurrentItem + fNumItemsToDraw/2.f;
-	int iFirstItemToDraw = std::ceil( fFirstItemToDraw );
-	int iLastItemToDraw = std::ceil( fLastItemToDraw );
+	int iFirstItemToDraw = static_cast<int>(std::ceil( fFirstItemToDraw ));
+	int iLastItemToDraw = static_cast<int>(std::ceil( fLastItemToDraw ));
 	if( !m_bLoop && !m_bWrap )
 	{
 		iFirstItemToDraw = std::clamp( iFirstItemToDraw, 0, m_iNumItems );
@@ -289,7 +289,7 @@ void ActorScroller::PositionItemsAndDrawPrimitives( bool bDrawPrimitives )
 		float fPosition = iItem - m_fCurrentItem;
 		int iIndex = i; // index into m_SubActors
 		if( m_bLoop || m_bWrap )
-			wrap( iIndex, m_SubActors.size() );
+			wrap( iIndex, static_cast<int>(m_SubActors.size()) );
 		else if( iIndex < 0 || iIndex >= (int)m_SubActors.size() )
 			continue;
